feat: adjusted chat ai
