+++ src/modules/azion-ai-chat/index.vue
@@ -3,115 +3,13 @@
     :style="{ 'background-color': 'var(--surface-section)' }"
     class="hidden flex-col w-full surface-border border-l h-[100ldh] transform translate-x-full transition-transform duration-300 ease-in-out sm:w-[500px] 2xl:w-[800px]"
   >
-    <AzionAiChatHeader>
-      <template #header-actions>
-        <PrimeButton
-          icon="pi pi-external-link"
-          outlined
-          class="surface-border h-8 w-8"
-          aria-label="Open a chat in new tab"
-          v-tooltip.bottom="'Open a chat in new tab'"
-          @click="openChatInNewTab"
-        />
-        <PrimeButton
-          icon="pi pi-pen-to-square"
-          outlined
-          class="surface-border h-8 w-8"
-          aria-label="New Chat"
-          v-tooltip.bottom="'New Chat'"
-          @click="handleClearChat"
-        />
-        <PrimeButton
-          icon="pi pi-times"
-          outlined
-          class="surface-border h-8 w-8"
-          aria-label="Close"
-          @click="azionAiChatStore.close()"
-        />
-      </template>
-    </AzionAiChatHeader>
-
-    <div
-      class="overflow-y-auto sticky top-12 h-[calc(100vh-5rem)] flex flex-col justify-between"
-      v-if="isChatMobile"
-    >
-      <AzionAiChat
-        :key="renderCount"
-        ref="azionAiChatRef"
-        :insert-deep-chat-styles="styleViewDeepChat"
-      />
-    </div>
-    <Sidebar
-      v-else
-      :visible="isChatAiOpen"
-      position="bottom"
-      headerContent="Copilot"
-      :show-close-icon="false"
-      :pt="{
-        root: { class: '!h-[100%] md:hidden flex' },
-        headerContent: { class: 'w-full' },
-        mask: { class: 'md:hidden flex' },
-        content: { class: '!p-0' }
-      }"
-    >
-      <template #header>
-        <div class="flex items-center justify-between">
-          <h2 class="flex items-center gap-2">
-            Copilot
-            <PrimeTag
-              class="ml-2"
-              value="Experimental"
-              v-tooltip.bottom="
-                'Copilot is in experimental mode and can give you some wrong answers. Please, always validate your answers.'
-              "
-              severity="info"
-            />
-          </h2>
-          <div class="gap-4 flex">
-            <PrimeButton
-              icon="pi pi-external-link"
-              outlined
-              class="surface-border h-8 w-8"
-              aria-label="Open a chat in new tab"
-              v-tooltip.bottom="'Open a chat in new tab'"
-              @click="openChatInNewTab"
-            />
-            <PrimeButton
-              icon="pi pi-pen-to-square"
-              outlined
-              class="surface-border h-8 w-8"
-              aria-label="New Chat"
-              v-tooltip.bottom="'New Chat'"
-              @click="handleClearChat"
-            />
-            <PrimeButton
-              icon="pi pi-times"
-              @click="azionAiChatStore.close()"
-              size="small"
-              class="flex-none surface-border text-sm w-8 h-8"
-              text
-            />
-          </div>
-        </div>
-      </template>
-
-      <div class="h-full w-full justify-between flex flex-col">
-        <AzionAiChat
-          :key="renderCount"
-          ref="azionAiChatMobileRef"
-        />
-      </div>
-    </Sidebar>
+    <azionAiChatDesktop />
   </div>
 </template>
 
 <script setup>
-  import PrimeTag from 'primevue/tag'
-  import Sidebar from 'primevue/sidebar'
-  import PrimeButton from 'primevue/button'
-  import AzionAiChatHeader from './azion-ai-chat-header-block.vue'
-  import AzionAiChat from './azion-ai-chat-block.vue'
-
+  import azionAiChatDesktop from './azion-ai-chat-desktop.vue'
+  import azionAiChatMobile from './azion-ai-chat-mobile.vue'
   import { useAzionAiChatStore } from '@/stores/azion-ai-chat-store'
   import { updateSessionId } from './services/make-session-id'
   import { computed, onMounted, onUnmounted, ref, watch } from 'vue'
@@ -119,7 +17,6 @@
   import { AZION_MESSAGE_TYPE } from '@modules/azion-ai-chat/directives/custom-ai-prompt'
   import { useRouter } from 'vue-router'
   import { windowOpen } from '@/helpers'
-  import { useWindowSize } from '@vueuse/core'
 
   defineOptions({
     name: 'azion-ai-chat-root-block'
@@ -128,12 +25,6 @@
   const azionAiChatMobileRef = ref(null)
   const renderCount = ref(1)
   const router = useRouter()
-  const { width } = useWindowSize()
-  const currentWidth = ref(width.value)
-  const SCREEN_BREAKPOINT_MD = 768
-  const styleViewDeepChat = {
-    border: 'none'
-  }
 
   onMounted(() => {
     window.addEventListener('message', aiCustomPromptListenerHandler)
@@ -142,10 +33,6 @@
 
   onUnmounted(() => {
     window.removeEventListener('message', aiCustomPromptListenerHandler)
-  })
-
-  const isChatMobile = computed(() => {
-    return currentWidth.value > SCREEN_BREAKPOINT_MD
   })
 
   const azionAiChatStore = useAzionAiChatStore()
@@ -186,12 +73,4 @@
     const url = `${window.location.origin}${router.resolve({ name: 'copilot' }).path}`
     windowOpen(url, '_blank')
   }
-
-  watch(
-    width,
-    () => {
-      currentWidth.value = width.value
-    },
-    { immediate: true }
-  )
 </script>
